requirements.txt
Pillow==10.0.0
PyQt6==6.5.2
openai>=1.0.0,<2.0.0
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
requests>=2.31.0
